+++ NEWS.md
@@ -1,5 +1,6 @@
 # simaerep 0.3.3
 - better data checks and error messages #19, #20, # 21
+- data check and repair by check_df_visit() became optionable
 
 # simaerep 0.3.2
 - added portfolio performance assessment